Perf log refactoring
This commit adds the ability to print out the performance logs
in between steps instead of just at the end of the simulation.
I'm adding this so we can get performance logging on longer
running simulations that may or may not finish on the cluster
before wall time runs out or a failure occurs.

closes #6300