Make test and modules code work with upcoming const interface changes.
Mainly this consists of fixing calls to coupledValue(),
coupledGradient(), value(), second(), secondTest(), secondPhi(), etc.
We can make this change *before* the changes are made at the framework
level to help ease the transistion.  All of the major apps and most of
the stork forks have already been updated as well.

Refs #6327.