Let FEProblem::addXYZ() functions update use_displaced_mesh.
All of the systems (Kernel, BoundaryCondition, UserObject, etc.) in
MOOSE are designed to work on the DisplacedMesh by default, and fall
back to working on the non-displaced mesh when there are no
displacement variables present.  This PR detects the latter case, and
changes the value of the use_displaced_mesh flag to false, to prevent
the user from seeing inconsistent values in these flags should they
happen to check them during a simulation.

Discovered while working on the bugfix in #6561.

A separate issue is the large amount of code duplication present in
the FEProblem::addXYZ() functions, but that is a topic for another PR.