Round flags that were stored as double
We were previously truncating, which was giving us erroneous answers
in cases where libMesh projection operation rounding error converted
an integer into integer-minus-epsilon.

This should be the first part of the fix for
https://github.com/idaholab/moose/issues/6621

After libMesh is modifed to avoid rounding error with constant
monomials, we might replace this with an assertion.