Permit multiple calls to registerObjects() and associateSyntax() ref #6757
There are multiple dependencies on other modules currently:
phase_field <- tensor_mechanics
solid_mechanics <- tensor_mechanics
xfem <- solid_mechanics

However, the dependent modules don't register the objects for the
apps they depend on to avoid double registration. This allows
them to do that, which permits tests that use capabilities from
a module that is depended upon to be included in the module
rather than having everything live in the combined module.