Fixing coordinate system check
The number of coordinate systems has to match the number of blocks.
If users specify blocks but no coordinate systems, we have to set the
coordinate systems for every block to XYZ (i.e. the default tvalue).
This typically happens when people use 'block' in [GlobalParams] and
use [Problem] block as well (typically with restart).

Closes #6765