PorousFlowMaterial1PhaseMD_Gaussian aded.
This is an exmple of how we can use nonlinear variables that aren't just porepressure or saturation: it uses log(mass-density).

Also, a derivative contribution to AdvectiveFlux that i forgot about previously
This meant adding properties to 1PhaseP_VG, 2PhasePP_VG and 2PhasePS.

Tests of the Jacobian for the new log(mass-density) material.
Pressure-pulse test for the new log(mass-density) material.