Bug fix in MD involving residual and jacobian
Also added bl02.i which is buckley-leverett in the MD formulation.
It gives the correct answer, but doesn't converge that well, and
for larger timesteps wanders off to very large P (at some stage
during the nonlinear iterative process) and gives a +inf
in the density, so then the simulations dies.