Level 1 Materials for PorousFlow
Including some Daniel changes

The Material architecture in the Porous Flow module is as follows.

(1) Level 1 Materials.  A user typically chooses one of these for their
particular simulation type, depending upon the overall type
of simulation.  For instance, there are Materials suitable for:
 (a) 1-phase saturated+unsaturated flow with porepressure as the primary nonlinear variable
 (b) 1-phase saturated+unsaturated flow with log(mass density) as the primary nonlinear variable
 (c) 2-phase flow with the two porepressures being the primary nonlinear variables
 (d) 2-phase flow with (porepressure0, saturation1) being the primary nonlinear variables
The purpose of these Level 1 Materials is to calculate, for each phase,
the porepressure, the saturation, gradients of these, and derivatives of
these wrt to the primary variables.  Also temperature, gradients of
temperature and derivatives of these wrt the primary variables.  Also
the mass fraction matrix, gradients and derivatives.  All these things
are calculated at the quadpoints and potentially the nodes depending on
whether we'll need nodal quantities in the mass-lumping and full-upwinding
in the Kernels.

(2) Level 2 Materials.  A user typically chooses many of these.
They are designed to get information from the Level 1 Materials and
calculate equation-of-state related things, such as fluid-phase density,
viscosity, relative permeability, and also solid-phase porosity.
They are coded so as to be independent of the Level 1 description, eg
a user can use the "water" density Material in both the 1-phase and
the 2-phase situation.  Typically these Level 2 Materials compute derivatives
of density, viscosity, etc, with respect to porepressures, saturations,
etc, independently of what the actual primary nonlinear variables are.

(3) Level 3 Materials.  These join the Level 2 Materials into std::vectors
for use in Kernels, and also join the derivative information of Level 2 Materials
into std::vectors that contain derivatives wrt the primary nonlinear variables
(ie, lots of chain rules here!).

Refs #6845