Enable users to plot relative norm of the difference in the solution
Refactored the computation of relative norm of the solution difference
- moved into NonlinearSystem. Reusing the function in Transient for both
Picard iterations and check for steady stete.  Improved the memory
allocation - the vector is allocated at the beginning and reused
throughout the simulation. Turning Transient::ss_relerr_norm into a
member variable with a better name.
Added a postprocessor for ploting the relative norm.
Added a test.

Refs #7118