Removing _axial_offset from RZ kernels
In order to use the RZ system provided by MOOSE, we need to remove the
axial offset parameter from our RZ kernels.  The axial offset has to be
handle by the geometry, i.e. we have to offset the elements physically
in the space.  Axial offset is now a parameter of the HeatStructure
component and the mesh generation algoritm respect it when the mesh is
generated.

Refs idaholab/relap-7#286