Adding basic structure for fluid properties module
There is a base class FluidProperties that inherits from a
GeneralUserObject. The next child class belwo it is
SinglePhaseFluidProperties than defines fluid properties for a single
phase like pressure, temperature, cp, cv, k, mu, c, etc.

Two basic equations of state are alos provided:
- stiffened gas
- ideal gas

There is a basic material that computes the properties:
FluidPropertiesMaterial.

The syntax to use this is:

[FluidProperties]
  [./air]
    type = IdealGasFluidProperties
    ...
  [../]
  [./water]
    type = StiffenedGasFluidProperties
    ...
  [../]
[]

Refs #6972