Add test of flow over Gaussian bump.
This test directory contains a sequence of uniformly-refined,
boundary-fitted Gmsh Q1 mesh files from @yidongxiainl.  The test uses
one of the coarse grids for speed, but it can use any of the files
listed in the input file to produce more accurate solutions.

This test also comes with a custom exodiff comparison file which use
absolute tolerances for the rhov and vel_y variables since they are of
very small magnitude in this problem and would otherwise produce
spurious rel_diffs in parallel.

Refs #3036.