Switch Navier-Stokes "bump" test to use bi-quadratic elements.
After discussing it with @yidongxiainl and running some more tests, we
found that the entropy error was not accurately estimated on the
bilinear grids, but converged at near the optimal rate on biquadratic
grids.  Note that we still solve with first-order variables, only the
geometry representation is updated by this change. This is a
well-known issue with continuous FEM approximations to the Euler
equations.

This PR also gets rid of the mesh files that were checked in,
replacing them with a Python script from @yidongxiainl that can be
used to generate any of the desired grids on the fly.

Refs #3036.