Fixed an issue with PETSc-3.7.x
Manually set a few entries of the solution vector to NAN to force the linear solver diverge and return  DIVERGED_NANORINF, when the system throws an exception.  There are two functions to determine convergence of the linear iterative solvers, one in PETSc and the other in MOOSE. In PETSc-3.7.X, the linear solver will not call the function in MOOSE any more once the PETSc function detects  divergence. it leaves the variable in MOOSE unreset,  and hence the old converged reason will be reused in the next linear solve. We fixed this issue by reseting the flag just right before the nonlinear solve every time.

Closes #6924