Add Action that adds InitialConditions.
To do this, we first added a new kind of InitialCondition that
requires a FluidProperties object.  Then, if the user provides an
initial pressure, temperature, and velocity vector, all of the
required Variable and AuxVariable values can be computed.

Changing the initial condition slightly caused the entropy_error
postprocessor value to change by O(1.e-15) in the first timestep which
caused a spurious exodiff, updated the bump_out.cmp file to fix this.

Refs #6972.