More extensive verification of TimeIntegrators for SCALAR ODEs.
This is a follow-up to the bug fixes in #6289.  The scalar.i input
file can now be used to verify that the different TimeIntegrator
methods have the correct asymptotic rates of convergence for SCALAR
ODE problems.

The run.py script runs this input file, and generates a convergence
plot in PDF format for the various methods.