Add tests of linear and nonlinear stiff ODEs.
Stiff ODEs frequently arise from FEM discretizations of multiphysics
PDEs, but we have not had a simple, good way of evaluating the
robustness of our TimeIntegrators for such problems before now.  This
test provides a way to make such an evaluation for both existing
TimeIntegrators and any new ones that we might add in the future.

We're just testing LStableDirk2 method by default.  The run_stiff.py
script can be used to generate results for other methods, and we have
used it to confirm that the expected order reduction occurs for some
of the higher-order methods.

Refs #5701.