Cutting timestep using MooseException is Impossible.
 This illustrates the Impossible error experienced by myself and others.  A Material calls

throw MooseExcecption("Something happened, dt needs to be reduced")

and MOOSE responds with Impossible.  Could someone take the time to fix this bug, please?

In this case, the material is ComputeLinearElasticStress_CUT_DT, that attempts to cut the timestep if stress(1, 1) > 1.
The test file that may be used to obtain the 'Impossible' error is push.i

Refs #3789