PorousFlow tests slightly reorganised.
Removing the Yaqi-Hacky exposed a problem in many of the tests.  Viz, the PorousFlowNodeNumber Material has to be defined before any other PorousFlowMaterial, since many of them use NodeNumber.  In particular, PorousFlowTemperature uses NodeNumber now that its Yaqi-Hacky is removed.

So, these are almost all trivial changes putting PorousFlowNodeNumber at the top of the list of Materials.

There are also a few other changes, to do with PETSc options that give better convergence.

Also, changed the dispersion tests to be the same as found in #7694 to avoid merge conflicts later on.

Fixes #7126