Add test of stagnation point (2D and RZ) incompressible flow.
These are classical "semi-analytical" solutions from boundary layer
theory which start from an inviscid (potential) flow solution and
modify it near a no-slip boundary to account for viscous effects.

The inviscid solutions (which are slightly different for 2D
vs. axisymmetric flow) are used as the "far field" boundary conditions
in the FEM simulation.  They are given by:

2D: u = K*x, v = -K*y
RZ: u_r = K*r, u_z = -2*K*z

where K is a given constant. It can be verified that both of these
solutions satisfy the divergence-free constraint in their respective
coordinate systems.  In the test harness, both configurations are run
via the same input file, with command line options specifying the
different kernels and boundary conditions where necessary.

Reproducing the analytical solution to the full Navier-Stokes
equations in this problem introduces several challenges:

1.) The far field boundary conditions, including the outlet, which is
    handled with a natural BC here, must be far enough away from the
    origin as to not have a strong effect on the flow there.
2.) The thickness of the boundary layer depends on the non-dimensional
    length scale sqrt(nu/K), where nu is the kinematic viscosity.  The
    boundary layer must be "thin" relative to the overall size of the
    domain in order for challenge 1 to be satisfied, but the thinner
    the boundary layer, the more mesh is required for an accurate
    solution.
3.) It was discovered that the outlet boundary cannot use the
    INSMomentumNoBCBC boundary condition, otherwise the problem is
    singular without a pressure pin.  The pressure pin gave fairly
    unsatisfying results for this problem, so the outlet is currently
    simply being handled with a natural BC (which is enough to
    constrain the non-trivial null-space of the problem) but seems to
    turn the flow a bit more than is desirable at the outlet.

There are a number of different ways of determining the accuracy of
the FEM solution with respect to the analytical solution for this
problem.  One of the simplest is to look at the "wall pressure"
i.e. the pressure field as a function of x (or r) starting at the
origin.  The analytical solution predicts that the wall pressure will
be given by

p(x) - p0 = -(1/2) * rho * K^2 * x^2

where "p0" is the stagnation pressure at the origin.  We observe
reasonably good convergence to this pressure field as we refine the
mesh and/or make the domain larger to reduce the effects of the
far-field boundary conditions on the flow. There is a
VectorPostprocessor in the input file which writes the wall pressure
to a CSV file for postprocessing, and can be used for verification
purposes.

Refs #3036.