Detect variables with different discretizations in the same NodalValueSampler.
The NodalValueSampler currently expects there to be a nodal value
available for each of its coupled variables at every node it
encounters.  This won't be the case when e.g. a FIRST, LAGRANGE
element is used on a mesh with quadratic elements.  In such cases we
can just skip sampling at the node in question, but in case someone
tries to construct a single NodalValueSampler with variables with
different discretizations, we don't have a good way of handling that
(e.g. we don't want to erroneously report 0 values for unsampled
variables) so we detect that case and throw an error.  The fix is to
use separate NodalValueSamplers for variables with different
discretizations. Also fix doco typo.