Add analytical solution verification test for Navier-Stokes.
Up to this point, we did not have a test verifying that the INS
kernels in MOOSE could reproduce an analytical solution; this commit
adds such a test. The application is "Jeffery-Hamel flow in a
wedge-shaped region," see also F. White, "Viscous Fluid Flow,"
pp. 168--172.

We also test the difference between applying the known solution as an
outlet boundary condition vs. using the "natural" BC at the outlet.
The former case converges to the analytical solution, while the latter
does not, despite the fact that both look "qualitatively" similar.
This is a good test case for demonstrating that the natural outflow BC
affects the entire solution field, and should not be expected to
reproduce a known solution in the general case.

* The included plot.py file shows that MOOSE obtains the optimal
  convergence rates for the velocity components in the L2-norm.
* Add Cubit journal file for creating the wedge geometry and input
  mesh file.
* Add WedgeFunction class to Navier-Stokes tests directory.  This
  function is only used by this specific test to read in the value of
  the semi-analytical known solution, linearly interpolating it if
  necessary.
* Add CSV file with semi-analytical exact solution data and Python
  script for generating the data for different configurations.

Refs #3036.