Rename classes and functions related to eigenstrain calculation ref #7815
Standardize on a naming scheme where we refer to these as 'eigenstrains' rather
than 'stress_free_strains'. Note that I am still leaving the material property
name untouched, as I am going to get rid of that in a later commit where the
user has to provide the property names.