Compute neighbor element volume on demand
MOOSE was computing the neighbor element volume even though it was not
really needed. The computation is quite expensive and shows up in
profiling.  If the volume is needed in a DG kernel, users can obtain it
by:

```
.h:
const Real & _neighbor_volume;

.C:
_neighbor_volume(getNeighborElemVolume())
```

This will tell MOOSE to compute the volume.

Refs #3661, #6788