Core changes
All PorousFlow Materials will now derive from PorousFlowMaterial that has a bool _nodal_material that gets set by all derived classes.
If the derived class sets _nodal_material to false then the class is a standard MOOSE Material, with Properties computed at the quadpoints.
If the derived class sets _nodal_material to true then the Properties are sized to _current_elem->n_nodes(), and the Properties may therefore be computed at the nodes of the element.

As explained in #8196, these new "nodal" Materials are necessary in PorousFlow because we lump and upwind in the Kernels.

Fixes #8196