Change diffusion_coefficient_name -> diffusion_coefficient.
Based on the discussion of #8254, the idea is that default value input
file syntax like:

specific_heat_name = 14.78

looks weird.

The goal is to drop the "_name" suffix from all the
MaterialPropertyName parameters, but this can't be done without a
deprecation period and application level changes for some names like
"density" because of their use in GlobalParams sections of input
files.