Add ComputeTemperatureDependentIsotropicElasticityTensor ref #8265
This deletes the ComputeVariableIsotropicElasticityTensor class
(which was set up for testing purposes) with a more general class
for temperature-dependent elastic constants prescribed by functions.

I converted the existing test over to using the new class just to
verify that it works, but will delete that test in a subsequent commit
that adds more thorough testing of this capability.