always call initQpStateful - remove no-override error
Previously, the error was important because stateful properties were not
properly zero-initialized.  So if materials' didn't override this
function, they would be uninitialized garbage.  But now, properties may
be stateful without the owning material knowing this - in these cases,
we still want to use the owner's initialization code (if provided).  So
we always want to call it.  Which means we can't throw an error if the
user didn't provide it (although it would be nice) - it would break too
much existing user code.