PQPlasticModel added.
This is a generalisation of the old ComputeCappedWeakPlaneStress to the case where arbitrary plastic models
are formulated in a 2-dimensional stress space, that i call p-q space.  PQPlasticModel does the return-map
process using a robust NR process that includes the possibility of sub-stepping the strain increment, as well
as a line search.

ComputeCappedWeakPlaneStress now derives from PQPlasticModel, making it much simpler.
ComputeCappedWeakInclinedPlaneStress derives from this, and so the const_cast stuff originally contained in this class is now not necessary.

No extra tests are included as the original tests were pretty exhaustive and this is just a refactor.

Fixes #8316