include self for supplied material property dep checking
Previously, material properties supplied by a material were not included
in the set of available properties a material could depend on.  This
prevented a material from calling getMaterialProperty* for any property
it declared/owned itself - because the checkDependMaterialsHelper func
would throw an error. This wasn't a big deal before properties could be
implicitly promoted to stateful.  Now that they can be - we plan on
deprecating the declareProperty[Old/Older] functions requiring materials
to use the getMaterialProperty[Old/Older] functions even for their own
properties.