Improved Performance Logging
This change makes the "Execution" section of the performance log much more dynamic.
Only active objects are logged and as more objects are added to a simulation, more
lines appear in the log to categorize them.

More importantly, this also now logs LINEAR and NONLINEAR UOs which was completely
missing before.

closes #8345