Fix to allow macros in bibtex files.
A new argument in configuration file can be supplied to point to where files of macro
definitions are as is the case with how the BISON bibtex folder is set-up.

Added test for this feature.

Added method that allows to temporarily change arguments in the configuration file
to test optional arguments that otherwise upset moosedocs.yml for MOOSE.

Refs #6699