Remove unused 'Leftover' bin for flow regime maps.
When designing the flow regime maps, we included a 'leftover' bin to flag using the original
closure while the TRACE closure was still being implemented.  This method has been
superseeded.  This update removes the bin on the 4 flow regime maps,  and fixes the
flow regime test input to not printout the nonexistent bin.
Ref idaholab/relap-7#19688