Add constant_on_elem parameter to Material.
When constant_on_elem is true, we call the user's computeQpProperties()
function only for qp 0, and then copy that value around to all the qps
for all the Properties declared for the Material.

A further refinement of this approach would completely avoid the
copying, but that requires deeper changes to possibly the MooseArray
class itself or possibly other aspects of the MaterialData/Storage
classes.

Slight rewording and fix to the documentation for "compute" parameter.

Refs #6979.