Skip repartitioning on oversampling meshes
When projecting from one distributed mesh to another, or when just
interpolating nodes because p refinement is for sissies or something,
we need the semilocal part of the source mesh to be a superset of the
local part of the destination mesh.

We can accomplish that here by simply turning off repartitioning,
since we got the destination mesh by cloning the source.

This gets us one step closer to fixing #1500