Divide by zero fix
Sometimes the memory logger will take a sample of a binary which has already
exited. This is treated as using zero bytes (which is correct after all).

However, if the binary's largest recorded memory usage is in fact 0 bytes,
which can happen if the binary started, performed its funciton, and exited all
before th ememory logger had a change to 'log it', the script will not be able
to scale the graph (bytes, megabytes, gigabytes, etc).

If 0 is detected, do not try to scale this number. Just plot the usage as a
literal zero bytes used.

Refs #8392