Bibtex hotfix to make Bison references not break.
For some reason, these few lines make Bison bibtex references not break when
"macro_files" is used in the MooseDocs.yml.  We think it has something
to do with other markdown references which aren't complete (don't call bibliography
and bibliography style) which causes some weird recursion.  But it works now.

Refs #6699