Add a no_timing flag to override timing flag
This aids us in testing. We want to verify that we can shutoff
timing through the normal input parameters but we sometimes pass
--timing to the app through the TestHarness. This flag is here
so that we don't have to fundamentally change the way we toggle
parameters, it simply takes precendence over --timing to allow
input file options to be tested.

closes #8493