Distributed MooseMesh::getBoundaryIDs(name)
We may not have all our boundary IDs locally, but userspace code is
probably expecting (in some cases definitely expecting) to be querying
the global boundary ID set, so let's use a global union to get it.

This fixes a dgkernels/dg_block_restrict.1D_test #1500 regression for
me with --distributed-mesh on 2 processors.