Distributed findImplicitGeometricCouplingEntries
If we're on a distributed mesh, we might not be able to access every
node, and we need to handle the cases where we can't.

I don't understand this code's downstream uses well enough to
*perfectly* parallelize it, so I'm doing some all-to-all communication
which isn't as efficient as it should be but which should ensure
correctness.

This fixes more test failures in #1500 for me