Don't mess with DistributedMesh allow_renumbering
If we want to turn that off via input file or via other code (e.g.
OversampleOutput) then we shouldn't override it here.  Some of the
MOOSE input file optiosn actually force users to refer to nodes by
number, and we want to allow those input files to work, even on a
DistributedMesh.