Distributed-compatible getNodeList()
I'm not sure this is a parallel_only() method, because
self-documenting self-asserting communication patterns apparently
haven't caught on outside of libMesh, so if we're on a distributed
mesh the best we can do here is return semilocal nodes and rely on
parallel user code to merge those sets if necessary.  That means that,
if we have an empty vector of nodes on a particular processor, we'd
better return a reference to an empty vector, not just throw an error.