Update glued contact tests ref #8636
These tests diffed with the change to volumetric locking becuase they
were ill-posed.  There was an initial gap between the two blocks, and
contact gets enforced at the point where the two blocks come in contact
during any iterations.  The location of that point can be very sensitive
to small changes in the behavior of the mechanics models, as happened
here.  I made a separate mesh for these models that has no initial gap
to make these more robust in the future.