Improving the way to obtain references to other components
A Component class now provides a new API to get references to other
components. This removes a lot of dynamic_casting that users had to do,
thus cleaning the code.

This also fixes the long present memory leak in PipeWithHeatStructure,
so we should have a clean valgrind target, finally.

Since we are obtainig const treferences a lot of API has to gain const
modifier.

Refs idaholab/relap-7#10, closes idaholab/relap-7#319