Cache and restore solutions on XFEM-affected nodes/elems ref #8763
We previously stored the dofs to copy the solution from, but sometimes
the dofs changed for existing nodes connected to XFEM elements.  This
caches the nl and aux solutions for all nodes and elements affected
by XFEM, which corrects that issue.  We also get the aux solution on
the new nodes/elems, which we didn't have before.