DependencyResolver comparator in depends multimap
This should complete the fixes that I *thought* I had finished in
PR #8674, thereby resolving #8659.  Now we have a consistent ordering
for *every* set produced by a resolver, not just every set except the
last.

This appears to fix intermittent #1500 failures for me.  As with the
previous #8659 fix it's hard to be certain, since there's always a
high chance that the previous broken code will work correctly via dumb
luck.