Modifications to FeatureFloodCount to support multiple IDs at point
This commit makes it so we can locate multiple grains at a single point
in a single pass. This is necessary for building up the IC because we
don't have a small number of disjoint features sitting on a small number
of OPs. We have potentially thousands of grains with a few active at
each point. Also gets the EBSD IC working

refs #8810