Launch all PBS jobs on HOLD
When using PBS options, place all launched jobs on hold as they
launch (qsub -h).

This fixes issues with pre-req tests completing before dependency
tests get launched. Because when this happens, the dependency test
will sit in the HOLD queue forever. Which will then require an
admin to release the job. I spoke to the torque guys, and this is
'operating as intended'.

Refs #9000